refactored note creation form to seperate module
