+++ part6/redux-anecdotes/src/components/AncedoteForm.jsx
@@ -1,0 +1,27 @@
+import { useDispatch } from "react-redux";
+import { createAnecdote } from "../reducers/anecdoteReducer";
+
+const AnecdoteForm = () => {
+  const dispatch = useDispatch();
+  const addAnecdote = (event) => {
+    event.preventDefault();
+    const contentTarget = event.target.content;
+    const action = createAnecdote(contentTarget.value);
+    dispatch(action);
+    contentTarget.value = "";
+  };
+
+  return (
+    <>
+      <h2>create new</h2>
+      <form onSubmit={addAnecdote}>
+        <div>
+          <input name="content" />
+        </div>
+        <button>create</button>
+      </form>
+    </>
+  );
+};
+
+export default AnecdoteForm;
